+++ launcher/minecraft/MinecraftInstance.cpp
@@ -1321,10 +1321,10 @@ QList<ResourceFolderModel*> MinecraftInstance::resourceLists()
     return { loaderModList(), coreModList(), nilModList(), resourcePackList(), texturePackList(), shaderPackList(), dataPackList() };
 }
 
-MultiWorldList* MinecraftInstance::worldList()
+WorldList* MinecraftInstance::worldList()
 {
     if (!m_world_list) {
-        m_world_list.reset(new MultiWorldList(QList{dynamic_cast<BaseInstance*>(this)}));
+        m_world_list.reset(new WorldList(QList{dynamic_cast<BaseInstance*>(this)}));
     }
     return m_world_list.get();
 }
