--- launcher/minecraft/MultiWorldList.cpp
@@ -1,504 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-/*
- *  Prism Launcher - Minecraft Launcher
- *  Copyright (C) 2022 Sefa Eyeoglu <contact@scrumplex.net>
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, version 3.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- * This file incorporates work covered by the following copyright and
- * permission notice:
- *
- *      Copyright 2013-2021 MultiMC Contributors
- *
- *      Licensed under the Apache License, Version 2.0 (the "License");
- *      you may not use this file except in compliance with the License.
- *      You may obtain a copy of the License at
- *
- *          http://www.apache.org/licenses/LICENSE-2.0
- *
- *      Unless required by applicable law or agreed to in writing, software
- *      distributed under the License is distributed on an "AS IS" BASIS,
- *      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *      See the License for the specific language governing permissions and
- *      limitations under the License.
- */
-
-#include "MultiWorldList.h"
-
-#include <FileSystem.h>
-#include <QDebug>
-#include <QDirIterator>
-#include <QFileSystemWatcher>
-#include <QMimeData>
-#include <QString>
-#include <QThreadPool>
-#include <QUrl>
-#include <QUuid>
-#include <Qt>
-
-#include "Application.h"
-#include "MinecraftInstance.h"
-#include "PackProfile.h"
-#include "icons/IconList.h"
-
-MultiWorldList::MultiWorldList(const QList<BaseInstance*>& instances) : QAbstractListModel(), allInstances(instances)
-{
-    for (BaseInstance* inst : allInstances) {
-        m_dirs.append(dynamic_cast<MinecraftInstance*>(inst)->worldDir());
-    }
-
-    for (QDir dir : m_dirs) {
-        FS::ensureFolderPathExists(dir.absolutePath());
-        dir.setFilter(QDir::Readable | QDir::NoDotAndDotDot | QDir::Files | QDir::Dirs);
-        dir.setSorting(QDir::Name | QDir::IgnoreCase | QDir::LocaleAware);
-    }
-
-    m_watcher = new QFileSystemWatcher(this);
-    m_isWatching = false;
-    connect(m_watcher, &QFileSystemWatcher::directoryChanged, this, &MultiWorldList::directoryChanged);
-}
-
-void MultiWorldList::startWatching()
-{
-    if (m_isWatching) {
-        return;
-    }
-    update();
-
-    m_isWatching = true;
-
-    for (const QDir& dir : m_dirs) {
-        if (m_watcher->addPath(dir.absolutePath())) {
-            qDebug() << "Started watching" << dir.absolutePath();
-        } else {
-            m_isWatching = false;
-            qDebug() << "Failed to start watching" << dir.absolutePath();
-        }
-    }
-}
-
-void MultiWorldList::stopWatching()
-{
-    if (!m_isWatching) {
-        return;
-    }
-
-    m_isWatching = false;
-
-    for (QDir dir : m_dirs) {
-        if (m_watcher->removePath(dir.absolutePath())) {
-            qDebug() << "Stopped watching" << dir.absolutePath();
-        } else {
-            m_isWatching = true;
-            qDebug() << "Failed to stop watching" << dir.absolutePath();
-        }
-    }
-}
-
-bool MultiWorldList::update()
-{
-    if (!isValid())
-        return false;
-
-    QList<InstanceWorld> newWorlds;
-
-    for (BaseInstance* inst : allInstances) {
-        QDir dir = dynamic_cast<MinecraftInstance*>(inst)->worldDir();
-        dir.refresh();
-        auto folderContents = dir.entryInfoList();
-        // if there are any untracked files...
-        for (QFileInfo entry : folderContents) {
-            if (!entry.isDir())
-                continue;
-
-            World w(entry);
-            if (w.isValid()) {
-                newWorlds.append(InstanceWorld(w, inst));
-            }
-        }
-    }
-    
-    beginResetModel();
-    m_worlds.swap(newWorlds);
-    endResetModel();
-    loadWorldsAsync();
-    return true;
-}
-
-void MultiWorldList::directoryChanged(QString)
-{
-    update();
-}
-
-bool MultiWorldList::isValid()
-{
-    bool valid = true;
-
-    for (const QDir& dir : m_dirs) {
-        if (!(dir.exists() && dir.isReadable())) {
-            valid = false;
-        }
-    }
-
-    return valid;
-}
-
-QList<QString> MultiWorldList::instDirPaths() const
-{
-    QList<QString> dirList;
-
-    for (BaseInstance const* instance : allInstances) {
-        dirList.append(QFileInfo(instance->instanceRoot()).absoluteFilePath());
-    }
-
-    return dirList;
-}
-
-bool MultiWorldList::deleteWorld(int index)
-{
-    if (index >= m_worlds.size() || index < 0)
-        return false;
-    World& m = m_worlds[index].world;
-    if (m.destroy()) {
-        beginRemoveRows(QModelIndex(), index, index);
-        m_worlds.removeAt(index);
-        endRemoveRows();
-        emit changed();
-        return true;
-    }
-    return false;
-}
-
-bool MultiWorldList::deleteWorlds(int first, int last)
-{
-    for (int i = first; i <= last; i++) {
-        World& m = m_worlds[i].world;
-        m.destroy();
-    }
-    beginRemoveRows(QModelIndex(), first, last);
-    m_worlds.erase(m_worlds.begin() + first, m_worlds.begin() + last + 1);
-    endRemoveRows();
-    emit changed();
-    return true;
-}
-
-bool MultiWorldList::resetIcon(int row)
-{
-    if (row >= m_worlds.size() || row < 0)
-        return false;
-    World& m = m_worlds[row].world;
-    if (m.resetIcon()) {
-        QModelIndex modelIndex = index(row, NameColumn);
-        emit dataChanged(modelIndex, modelIndex, { MultiWorldList::IconFileRole });
-        return true;
-    }
-    return false;
-}
-
-int MultiWorldList::columnCount(const QModelIndex& parent) const
-{
-    return parent.isValid() ? 0 : 7;
-}
-
-QVariant MultiWorldList::data(const QModelIndex& index, int role) const
-{
-    if (!index.isValid())
-        return QVariant();
-
-    int row = index.row();
-    int column = index.column();
-
-    if (row < 0 || row >= m_worlds.size())
-        return QVariant();
-
-    QLocale locale;
-
-    const auto& instanceWorld = m_worlds[row];
-    switch (role) {
-        case Qt::DisplayRole:
-            switch (column) {
-                case NameColumn:
-                    return instanceWorld.world.name();
-
-                case InstanceColumn:
-                    return instanceWorld.instance->name();
-
-                case VersionColumn:
-                    static_cast<MinecraftInstance*>(instanceWorld.instance)->getPackProfile()->reload(Net::Mode::Online);
-                    return static_cast<MinecraftInstance*>(instanceWorld.instance)->getPackProfile()->getComponentVersion("net.minecraft");
-
-                case GameModeColumn:
-                    return instanceWorld.world.gameType().toTranslatedString();
-
-                case LastPlayedColumn:
-                    return instanceWorld.world.lastPlayed();
-
-                case SizeColumn:
-                    return locale.formattedDataSize(instanceWorld.world.bytes());
-
-                case InfoColumn:
-                    if (instanceWorld.world.isSymLinkUnder(QFileInfo(instanceWorld.instance->instanceRoot()).absoluteFilePath())) {
-                        return tr("This world is symbolically linked from elsewhere.");
-                    }
-                    if (instanceWorld.world.isMoreThanOneHardLink()) {
-                        return tr("\nThis world is hard linked elsewhere.");
-                    }
-                    return "";
-                default:
-                    return QVariant();
-            }
-
-        case Qt::UserRole:
-            if (column == SizeColumn)
-                return QVariant::fromValue<qlonglong>(instanceWorld.world.bytes());
-            return data(index, Qt::DisplayRole);
-
-        case Qt::ToolTipRole: {
-            if (column == InfoColumn) {
-                if (instanceWorld.world.isSymLinkUnder(QFileInfo(instanceWorld.instance->instanceRoot()).absoluteFilePath())) {
-                    return tr("Warning: This world is symbolically linked from elsewhere. Editing it will also change the original."
-                        "\nCanonical Path: %1")
-                        .arg(instanceWorld.world.canonicalFilePath());
-                }
-                if (instanceWorld.world.isMoreThanOneHardLink()) {
-                    return tr("Warning: This world is hard linked elsewhere. Editing it will also change the original.");
-                }
-            }
-            return instanceWorld.world.folderName();
-        }
-        case ObjectRole: {
-            return QVariant::fromValue<void*>((void*)&instanceWorld);
-        }
-        case FolderRole: {
-            return QDir::toNativeSeparators(QDir(dynamic_cast<MinecraftInstance*>(instanceWorld.instance)->worldDir()).absoluteFilePath(instanceWorld.world.folderName()));
-        }
-        case SeedRole: {
-            return QVariant::fromValue<qlonglong>(instanceWorld.world.seed());
-        }
-        case NameRole: {
-            return instanceWorld.world.name();
-        }
-        case LastPlayedRole: {
-            return instanceWorld.world.lastPlayed();
-        }
-        case SizeRole: {
-            return QVariant::fromValue<qlonglong>(instanceWorld.world.bytes());
-        }
-        case IconFileRole: {
-            return instanceWorld.world.iconFile();
-        }
-        case InstanceIconFileRole: {
-            return APPLICATION->icons()->getIcon(instanceWorld.instance->iconKey());
-        }
-        default:
-            return QVariant();
-    }
-}
-
-QVariant MultiWorldList::headerData(int section, [[maybe_unused]] Qt::Orientation orientation, int role) const
-{
-    switch (role) {
-        case Qt::DisplayRole:
-            switch (section) {
-                case NameColumn:
-                    return tr("Name");
-                case InstanceColumn:
-                    return tr("Instance");
-                case VersionColumn:
-                    return tr("Version");
-                case GameModeColumn:
-                    return tr("Game Mode");
-                case LastPlayedColumn:
-                    return tr("Last Played");
-                case SizeColumn:
-                    //: World size on disk
-                    return tr("Size");
-                case InfoColumn:
-                    //: special warnings?
-                    return tr("Info");
-                default:
-                    return QVariant();
-            }
-
-        case Qt::ToolTipRole:
-            switch (section) {
-                case NameColumn:
-                    return tr("The name of the world.");
-                case InstanceColumn:
-                    return tr("The instance the world belongs to.");
-                case VersionColumn:
-                    return tr("The Minecraft version of the world.");
-                case GameModeColumn:
-                    return tr("Game mode of the world.");
-                case LastPlayedColumn:
-                    return tr("Date and time the world was last played.");
-                case SizeColumn:
-                    return tr("Size of the world on disk.");
-                case InfoColumn:
-                    return tr("Information and warnings about the world.");
-                default:
-                    return QVariant();
-            }
-        default:
-            return QVariant();
-    }
-}
-
-QStringList MultiWorldList::mimeTypes() const
-{
-    QStringList types;
-    types << "text/uri-list";
-    return types;
-}
-
-QMimeData* MultiWorldList::mimeData(const QModelIndexList& indexes) const
-{
-    QList<QUrl> urls;
-
-    for (auto idx : indexes) {
-        if (idx.column() != 0)
-            continue;
-
-        int row = idx.row();
-        if (row < 0 || row >= this->m_worlds.size())
-            continue;
-
-        const World& world = m_worlds[row].world;
-
-        if (!world.isValid() || !world.isOnFS())
-            continue;
-
-        QString worldPath = world.container().absoluteFilePath();
-        qDebug() << worldPath;
-        urls.append(QUrl::fromLocalFile(worldPath));
-    }
-
-    auto result = new QMimeData();
-    result->setUrls(urls);
-    return result;
-}
-
-Qt::ItemFlags MultiWorldList::flags(const QModelIndex& index) const
-{
-    Qt::ItemFlags defaultFlags = QAbstractListModel::flags(index);
-    if (index.isValid())
-        return Qt::ItemIsUserCheckable | Qt::ItemIsDragEnabled | Qt::ItemIsDropEnabled | defaultFlags;
-    else
-        return Qt::ItemIsDropEnabled | defaultFlags;
-}
-
-Qt::DropActions MultiWorldList::supportedDragActions() const
-{
-    // move to other mod lists or VOID
-    return Qt::MoveAction;
-}
-
-Qt::DropActions MultiWorldList::supportedDropActions() const
-{
-    // copy from outside, move from within and other mod lists
-    return Qt::CopyAction | Qt::MoveAction;
-}
-
-void MultiWorldList::installWorld(BaseInstance* instance, const QFileInfo& filename)
-{
-    qDebug() << "installing:" << filename.absoluteFilePath();
-    World w(filename);
-    if (!w.isValid()) {
-        return;
-    }
-    w.install(QDir(dynamic_cast<MinecraftInstance*>(instance)->worldDir()).absolutePath());
-    update();
-}
-
-void MultiWorldList::installWorld(const QFileInfo& filename)
-{
-    if (allInstances.size() == 1) {
-        installWorld(allInstances[0], filename);
-    } else {
-        qDebug() << "world could not be installed, instance was not specified";
-    }
-}
-
-bool MultiWorldList::dropMimeData(const QMimeData* data,
-                             Qt::DropAction action,
-                             [[maybe_unused]] int row,
-                             [[maybe_unused]] int column,
-                             [[maybe_unused]] const QModelIndex& parent)
-{
-    if (action == Qt::IgnoreAction)
-        return true;
-    // check if the action is supported
-    if (!data || !(action & supportedDropActions()))
-        return false;
-    // files dropped from outside?
-    if (data->hasUrls()) {
-        bool was_watching = m_isWatching;
-        if (was_watching)
-            stopWatching();
-        auto urls = data->urls();
-        for (auto url : urls) {
-            // only local files may be dropped...
-            if (!url.isLocalFile())
-                continue;
-            QString filename = url.toLocalFile();
-
-            QFileInfo worldInfo(filename);
-
-            emit fileDropped(worldInfo);
-        }
-        if (was_watching)
-            startWatching();
-        return true;
-    }
-    return false;
-}
-
-int64_t MultiWorldList::calculateWorldSize(const QFileInfo& file)
-{
-    if (file.isFile() && file.suffix() == "zip") {
-        return file.size();
-    } else if (file.isDir()) {
-        QDirIterator it(file.absoluteFilePath(), QDir::Files, QDirIterator::Subdirectories);
-        int64_t total = 0;
-        while (it.hasNext()) {
-            it.next();
-            total += it.fileInfo().size();
-        }
-        return total;
-    }
-    return -1;
-}
-
-void MultiWorldList::loadWorldsAsync()
-{
-    for (int i = 0; i < m_worlds.size(); ++i) {
-        auto file = m_worlds.at(i).world.container();
-        int row = i;
-        QThreadPool::globalInstance()->start([this, file, row]() mutable {
-            auto size = calculateWorldSize(file);
-
-            QMetaObject::invokeMethod(
-                this,
-                [this, size, row, file]() {
-                    if (row < m_worlds.size() && m_worlds[row].world.container() == file) {
-                        m_worlds[row].world.setSize(size);
-
-                        // Notify views
-                        QModelIndex modelIndex = index(row, SizeColumn);
-                        emit dataChanged(modelIndex, modelIndex, { SizeRole });
-                    }
-                },
-                Qt::QueuedConnection);
-        });
-    }
-}
