--- launcher/ui/MultiWorldListPage.ui
@@ -1,196 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>MultiWorldListPage</class>
- <widget class="QMainWindow" name="MultiWorldListPage">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>800</width>
-    <height>600</height>
-   </rect>
-  </property>
-  <widget class="QWidget" name="centralwidget">
-   <layout class="QHBoxLayout" name="horizontalLayout">
-    <property name="leftMargin">
-     <number>0</number>
-    </property>
-    <property name="topMargin">
-     <number>0</number>
-    </property>
-    <property name="rightMargin">
-     <number>0</number>
-    </property>
-    <property name="bottomMargin">
-     <number>0</number>
-    </property>
-    <item>
-     <layout class="QVBoxLayout" name="verticalLayout">
-      <item>
-       <widget class="QTreeView" name="worldTreeView">
-        <property name="sizePolicy">
-         <sizepolicy hsizetype="Expanding" vsizetype="Expanding">
-          <horstretch>0</horstretch>
-          <verstretch>0</verstretch>
-         </sizepolicy>
-        </property>
-        <property name="acceptDrops">
-         <bool>true</bool>
-        </property>
-        <property name="dragDropMode">
-         <enum>QAbstractItemView::DragDrop</enum>
-        </property>
-        <property name="alternatingRowColors">
-         <bool>true</bool>
-        </property>
-        <property name="rootIsDecorated">
-         <bool>false</bool>
-        </property>
-        <property name="itemsExpandable">
-         <bool>false</bool>
-        </property>
-        <property name="sortingEnabled">
-         <bool>true</bool>
-        </property>
-        <property name="allColumnsShowFocus">
-         <bool>true</bool>
-        </property>
-        <property name="verticalScrollMode">
-         <enum>QAbstractItemView::ScrollPerPixel</enum>
-        </property>
-        <attribute name="headerStretchLastSection">
-         <bool>false</bool>
-        </attribute>
-       </widget>
-      </item>
-      <item>
-       <widget class="QLineEdit" name="filterEdit">
-        <property name="placeholderText">
-         <string>Search</string>
-        </property>
-       </widget>
-      </item>
-      </layout>
-    </item>
-   </layout>
-  </widget>
-  <widget class="WideBar" name="toolBar">
-   <property name="windowTitle">
-    <string>Actions</string>
-   </property>
-   <property name="allowedAreas">
-    <set>Qt::LeftToolBarArea|Qt::RightToolBarArea</set>
-   </property>
-   <property name="toolButtonStyle">
-    <enum>Qt::ToolButtonTextOnly</enum>
-   </property>
-   <property name="floatable">
-    <bool>false</bool>
-   </property>
-   <attribute name="toolBarArea">
-    <enum>RightToolBarArea</enum>
-   </attribute>
-   <attribute name="toolBarBreak">
-    <bool>false</bool>
-   </attribute>
-   <addaction name="actionAdd"/>
-   <addaction name="separator"/>
-   <addaction name="actionJoin"/>
-   <addaction name="actionJoin_Offline"/>
-   <addaction name="actionRename"/>
-   <addaction name="actionCopy"/>
-   <addaction name="actionRemove"/>
-   <addaction name="actionInstance_Settings"/>
-   <addaction name="actionMCEdit"/>
-   <addaction name="actionData_Packs"/>
-   <addaction name="actionReset_Icon"/>
-   <addaction name="separator"/>
-   <addaction name="actionCopy_Seed"/>
-   <addaction name="actionRefresh"/>
-   <addaction name="actionView_Folder"/>
-  </widget>
-  <action name="actionAdd">
-   <property name="text">
-    <string>Add</string>
-   </property>
-  </action>
-  <action name="actionJoin">
-   <property name="text">
-    <string>Join</string>
-   </property>
-  </action>
-  <action name="actionJoin_Offline">
-   <property name="text">
-    <string>Join Offline</string>
-   </property>
-  </action>
-  <action name="actionRename">
-   <property name="text">
-    <string>Rename</string>
-   </property>
-  </action>
-  <action name="actionCopy">
-   <property name="text">
-    <string>Copy</string>
-   </property>
-  </action>
-  <action name="actionRemove">
-   <property name="text">
-    <string>Delete</string>
-   </property>
-  </action>
-  <action name="actionMCEdit">
-   <property name="text">
-    <string>MCEdit</string>
-   </property>
-  </action>
-  <action name="actionCopy_Seed">
-   <property name="text">
-    <string>Copy Seed</string>
-   </property>
-  </action>
-  <action name="actionRefresh">
-   <property name="text">
-    <string>Refresh</string>
-   </property>
-  </action>
-  <action name="actionView_Folder">
-   <property name="text">
-    <string>View Folder</string>
-   </property>
-  </action>
-  <action name="actionReset_Icon">
-   <property name="text">
-    <string>Reset Icon</string>
-   </property>
-   <property name="toolTip">
-    <string>Remove world icon to make the game re-generate it on next load.</string>
-   </property>
-  </action>
-  <action name="actionData_Packs">
-   <property name="text">
-    <string>Data Packs</string>
-   </property>
-   <property name="toolTip">
-    <string>Manage data packs inside the world.</string>
-   </property>
-  </action>
-  <action name="actionInstance_Settings">
-   <property name="text">
-    <string>View Instance</string>
-   </property>
-   <property name="toolTip">
-    <string>View the instance this world belongs to.</string>
-   </property>
-  </action>
- </widget>
- <customwidgets>
-  <customwidget>
-   <class>WideBar</class>
-   <extends>QToolBar</extends>
-   <header>ui/widgets/WideBar.h</header>
-  </customwidget>
- </customwidgets>
- <resources/>
- <connections/>
-</ui>
